Final commit for HW_05
