+++ src/config/config.js
@@ -5,15 +5,15 @@ export const config = {
     'callbackURL': 'http://localhost:8080/auth/facebook/callback',
   },
   
-  'twitterAuth' : {
-    'consumerKey'        : 'tD06IBLf6nO9U2SMybgMdHyC6',
-    'consumerSecret'     : 'nrmBO0fIOgn3woXkkcfPsMeDHCv2hIBFhdwCoArN6GhX5GrIIA',
-    'callbackURL'        : 'http://localhost:8080/auth/twitter/callback'
+  'twitterAuth': {
+    'consumerKey': 'tD06IBLf6nO9U2SMybgMdHyC6',
+    'consumerSecret': 'nrmBO0fIOgn3woXkkcfPsMeDHCv2hIBFhdwCoArN6GhX5GrIIA',
+    'callbackURL': 'http://localhost:8080/auth/twitter/callback'
   },
   
-  'googleAuth' : {
-    'clientID'         : '427037420027-50dlbv7pmhhhfi6p7i43lgf1tpt2rnim.apps.googleusercontent.com',
-    'clientSecret'     : 'OrsUKzyrjcdcK5e7clN1Fe22',
-    'callbackURL'      : 'http://localhost:8080/auth/google/callback'
+  'googleAuth': {
+    'clientID': '427037420027-50dlbv7pmhhhfi6p7i43lgf1tpt2rnim.apps.googleusercontent.com',
+    'clientSecret': 'OrsUKzyrjcdcK5e7clN1Fe22',
+    'callbackURL': 'http://localhost:8080/auth/google/callback'
   }
 };
