+++ src/data/auth_data.json
@@ -18,8 +18,4 @@
     "password": "12345",
     "test": "google"
   }
-  
-  
-  
-  
 ]
