+++ src/middlewares/check-token
@@ -1,24 +1,15 @@
 import jwt from 'jsonwebtoken';
+
 /**
  * @description
- * Middleware for cookie parsing in request.
- * Parse Cookie header and populate `req.parsedCookies`
- * with an object keyed and value by the cookie data.
+ * Middleware for checking token in request.
  *
  * @param {Request} req
  * @param {http.ServerResponse} res
  * @param {Function} next
  */
-/**
- *
- * @param req
- * @param res
- * @param next
- * @return {*}
- */
 export const checkToken = (req, res, next) => {
   const token = req.headers['x-access-token'];
-  
   if (token) {
     jwt.verify(token, 'secretKey', (err) => {
       if (err) {
