--- src/strategies/facebook
@@ -1,27 +1,0 @@
-import passportFacebook from 'passport-facebook';
-
-import authData from '../data/auth_data.json';
-import { config } from '../config/config';
-
-const Strategy = passportFacebook.Strategy;
-
-export const facebookStrategy = new Strategy({
-  clientID: config.facebookAuth.clientID,
-  clientSecret: config.facebookAuth.clientSecret,
-  callbackURL: config.facebookAuth.callbackURL
-},
-(accessToken, refreshToken, profile, done) => {
-  process.nextTick(() => {
-    let user = null;
-    authData.forEach(data => {
-      if (data.login === profile.displayName) {
-        user = data;
-      }
-    });
-    if (!user) {
-      done(null, false, 'There are no such user in database');
-    } else {
-      done(null, user);
-    }
-  });
-});
